backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mistralai>=1.2.0
psycopg[binary]>=3.2.0
pgvector>=0.3.0
python-dotenv>=1.0.0
